Spikes adjustments
Added an enum to spikes.cs to denote the UP or DOWN state.
Adjusted the animation to make the down period slightly longer than it
had been.
The animation changes the UP and DOWN state of the spike enum according
to the state of the sprite.
The OnCollisionEnter2D of spikes.cs will respond differently depending
on UP or DOWN.